--- routes/protected/howtoplay_auth.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-/* GET page. */
-router.get('/', function(req, res, next) {
-    res.render("authenticated/howtoplay_auth", {title:"How to Play Page (Auth)"});
-  });
-
-
-module.exports = router;
