--- routes/protected/lobby.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render("authenticated/lobby", {title:"Lobby Page"});
-});
-
-
-module.exports = router;
