+++ routes/public/index.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
+
 /* GET home page. */
 router.get('/', function(req, res, next) {
   res.render("unauthenticated/home", {title:"HOME PAGE"});
@@ -25,6 +26,10 @@ router.get('/howtoplay', function (req, res, next) {
 router.get('/about', function (req, res, next) {
   res.render("unauthenticated/about", {title:"About Page"});
 });
+
+
+
+
 
 module.exports = router;
